revert changes to TypeHint
change suite name

move method to class
