+++ function-aws-custom-runtime/src/main/java/io/micronaut/function/aws/runtime/AbstractMicronautLambdaRuntime.java
@@ -28,6 +28,7 @@ import io.micronaut.context.env.CommandLinePropertySource;
 import io.micronaut.context.exceptions.ConfigurationException;
 import io.micronaut.core.annotation.NonNull;
 import io.micronaut.core.annotation.Nullable;
+import io.micronaut.core.annotation.TypeHint;
 import io.micronaut.core.beans.BeanIntrospection;
 import io.micronaut.core.cli.CommandLine;
 import io.micronaut.core.reflect.ClassUtils;
@@ -56,6 +57,13 @@ import java.util.Map;
 import java.util.Optional;
 import java.util.function.Predicate;
 
+import static io.micronaut.core.annotation.TypeHint.AccessType.ALL_DECLARED_CONSTRUCTORS;
+import static io.micronaut.core.annotation.TypeHint.AccessType.ALL_DECLARED_FIELDS;
+import static io.micronaut.core.annotation.TypeHint.AccessType.ALL_DECLARED_METHODS;
+import static io.micronaut.core.annotation.TypeHint.AccessType.ALL_PUBLIC;
+import static io.micronaut.core.annotation.TypeHint.AccessType.ALL_PUBLIC_CONSTRUCTORS;
+import static io.micronaut.core.annotation.TypeHint.AccessType.ALL_PUBLIC_FIELDS;
+import static io.micronaut.core.annotation.TypeHint.AccessType.ALL_PUBLIC_METHODS;
 import static io.micronaut.http.HttpHeaders.USER_AGENT;
 
 /**
@@ -70,10 +78,40 @@ import static io.micronaut.http.HttpHeaders.USER_AGENT;
  * @author sdelamo
  * @since 2.0.0
  */
-
+@TypeHint(
+    accessType = {
+        ALL_PUBLIC,
+        ALL_DECLARED_CONSTRUCTORS,
+        ALL_PUBLIC_CONSTRUCTORS,
+        ALL_DECLARED_METHODS,
+        ALL_DECLARED_FIELDS,
+        ALL_PUBLIC_METHODS,
+        ALL_PUBLIC_FIELDS
+    },
+    value = {
+        com.amazonaws.services.lambda.runtime.events.ApplicationLoadBalancerRequestEvent.class,
+        com.amazonaws.services.lambda.runtime.events.ApplicationLoadBalancerResponseEvent.class,
+        com.amazonaws.services.lambda.runtime.events.APIGatewayProxyRequestEvent.class,
+        com.amazonaws.services.lambda.runtime.events.APIGatewayProxyRequestEvent.ProxyRequestContext.class,
+        com.amazonaws.services.lambda.runtime.events.APIGatewayProxyRequestEvent.RequestIdentity.class,
+        com.amazonaws.services.lambda.runtime.events.APIGatewayProxyResponseEvent.class,
+        com.amazonaws.services.lambda.runtime.events.ScheduledEvent.class,
+        com.amazonaws.services.lambda.runtime.events.APIGatewayV2ProxyRequestEvent.class,
+        com.amazonaws.services.lambda.runtime.events.APIGatewayV2ProxyResponseEvent.class,
+        com.amazonaws.services.lambda.runtime.events.CloudFrontEvent.class,
+        com.amazonaws.services.lambda.runtime.events.CloudWatchLogsEvent.class,
+        com.amazonaws.services.lambda.runtime.events.CodeCommitEvent.class,
+        com.amazonaws.services.lambda.runtime.events.CognitoEvent.class,
+        com.amazonaws.services.lambda.runtime.events.ConfigEvent.class,
+        com.amazonaws.services.lambda.runtime.events.IoTButtonEvent.class,
+        com.amazonaws.services.lambda.runtime.events.LexEvent.class,
+        com.amazonaws.services.lambda.runtime.events.SNSEvent.class,
+        com.amazonaws.services.lambda.runtime.events.SQSEvent.class
+    }
+)
 @SuppressWarnings("java:S119") // More descriptive generics are better here
 public abstract class AbstractMicronautLambdaRuntime<RequestType, ResponseType, HandlerRequestType, HandlerResponseType>
-        implements ApplicationContextProvider, AwsLambdaRuntimeApi {
+    implements ApplicationContextProvider, AwsLambdaRuntimeApi {
     @Nullable
     protected String userAgent;
 
@@ -147,8 +185,8 @@ public abstract class AbstractMicronautLambdaRuntime<RequestType, ResponseType, 
     public ApplicationContextBuilder createApplicationContextBuilderWithArgs(String... args) {
         CommandLine commandLine = CommandLine.parse(args);
         return ApplicationContext.builder()
-                .environments(MicronautLambdaContext.ENVIRONMENT_LAMBDA)
-                .propertySources(new CommandLinePropertySource(commandLine));
+            .environments(MicronautLambdaContext.ENVIRONMENT_LAMBDA)
+            .propertySources(new CommandLinePropertySource(commandLine));
     }
 
     /**
@@ -293,16 +331,7 @@ public abstract class AbstractMicronautLambdaRuntime<RequestType, ResponseType, 
             log(LogLevel.TRACE, "request of type APIGatewayV2HTTPEvent\n");
             String content = apiGatewayV2HTTPEvent.getBody();
             return valueFromContent(content, handlerRequestType);
-        } else if (request instanceof com.amazonaws.services.lambda.runtime.events.APIGatewayProxyRequestEvent apiGatewayProxyRequestEvent) {
-            log(LogLevel.TRACE, "request of type APIGatewayProxyRequestEvent");
-            String content = apiGatewayProxyRequestEvent.getBody();
-            return valueFromContent(content, handlerRequestType);
-        } else if (request instanceof com.amazonaws.services.lambda.runtime.events.APIGatewayV2HTTPEvent apiGatewayV2HTTPEvent) {
-            log(LogLevel.TRACE, "request of type APIGatewayV2HTTPEvent\n");
-            String content = apiGatewayV2HTTPEvent.getBody();
-            return valueFromContent(content, handlerRequestType);
         }
-
 
         log(LogLevel.TRACE, "createHandlerRequest return null\n");
         return null;
@@ -331,9 +360,9 @@ public abstract class AbstractMicronautLambdaRuntime<RequestType, ResponseType, 
             config.setReadTimeout(null);
             config.setConnectTimeout(null);
             final HttpClient endpointClient = applicationContext.createBean(
-                    HttpClient.class,
-                    runtimeApiURL,
-                    config);
+                HttpClient.class,
+                runtimeApiURL,
+                config);
             final BlockingHttpClient blockingHttpClient = endpointClient.toBlocking();
             try {
                 while (loopUntil.test(runtimeApiURL)) {
@@ -552,8 +581,8 @@ public abstract class AbstractMicronautLambdaRuntime<RequestType, ResponseType, 
     @SuppressWarnings("rawtypes")
     private Class initTypeArgument(int index) {
         final Class[] args = GenericTypeUtils.resolveSuperTypeGenericArguments(
-                getClass(),
-                AbstractMicronautLambdaRuntime.class
+            getClass(),
+            AbstractMicronautLambdaRuntime.class
         );
         if (ArrayUtils.isNotEmpty(args) && args.length > index) {
             return args[index];
