+++ test-suite-aws-lambda-events-serde/src/test/java/io/micronaut/aws/lambda/events/AwsLambdaEventsSerdeSuite.java
@@ -6,6 +6,6 @@ import org.junit.platform.suite.api.SuiteDisplayName;
 
 @Suite
 @SelectPackages("io.micronaut.aws.lambda.events.tests")
-@SuiteDisplayName("AWS Lambda Events TCK for JacksonDatabind")
+@SuiteDisplayName("AWS Lambda Events TCK for Serde")
 public class AwsLambdaEventsSerdeSuite {
 }
